Prepwork for dealing with the common encrypted ports.
In order to maintain compatibility with existing deployed nodes, known common encrypted ports will NOT be blocked by default.

Users will receive a message during first setup encouraging them to review the rules that apply to how they intend to use their node and that laws very by country and frequency.

A package blockknownencrypted has been created in changeset:123a521df2b63ba1c5bdd6ad94ac402b86394579/bbhn_packages to be used in blocking known encrypted ports if the user feels it is necessary.

As developers we are not stating an opinion as to what the rules say or do not say in relationship to the traffic this deals with.  Each user will need to make their own determination of the rules.

This has been the current case since day one.

New file fwinfo page shows if the package is installed AND displays the active firewall rules at the time of the access.  This allows future grown to help test (because of the adhoc nature of the mesh) if packet filtering is the cause of a connection not working.

see BBHN->ticket:3