bugfix: Links on status page use midX domains which are not resolveable on host they reference.
mid hostname entries are only created on remote nodes and are not created on the local node (by design of nameservice module)

In order to allow links on remote nodes olsr status screens and potentionaly other locations to resolve correctly we must add a hostname for the dtdlink interface into the name service beacons.
In addtion we need to ensure the UI looks at these links instead.

This feature will need to be looked at for long in the mesh status screen to find a better method to display these connections.

ref BBHN->ticket:47