bugfix: non mesh-gw nodes may use a less than optimum route to internet.
When a non mesh-gw node has a route via the WAN interface and through a mesh-gw it will choose the mesh-gw instead of the local WAN connection.and

We change the routing list to prefer the WAN link.

When the WAN is static configured the user will need to disable the WAN interface for the mesh internet to be chosen.

When the WAN is configured to dynamic the loss of a DHCP lease will allow the node to chose the remote mesh internet.

The simple solution would be to change the WAN to disabled and reboot to ensure a remote node is used instead if the local internet fails.

Thanks to Joe AE6XE and Clint AE5CA for pointing out this scenario.