bugfix: olsrd-watchdog false triggers a reboot.
Using the method suggested by Henning Rogge in 2014

{{{
1) remove file
2) wait 2 intervals
3) if file exists, go to 1)
4) restart olsr, then go to 1)
}}}

With the exception that we currently wait more than 2 intervals.

Even if the script hits a collision at the same time the file is being written the file will already exist from one of the previous writes or the current write.

Removing a file in use is safe as once the FD closes the file contents are fully released.

Additional Advantages:
* Should be more efficient to check if file exists instead of reading the file and comparing date stamps.
* Removing the Perl code moves us one step closer to retiring Perl in the future.