bugfix: nodes sometimes show up as mid1
Nodes sometimes show up as a mid1 entry when dtdlink and RF connected.

Sometimes the nodes will get the eth0 ip because wlan0 is not up yet.

Under Barrier Breaker the MainIP option is now supported as part of the uci system which allows us to restore using this setting.

We can now force the IP address that is claimed (wlan0) for the primary IP of the node.