feature: migrate scripts to use get_interface
nvram-setup may be the only script we have to worry about as it may call configs for interfaces that do not exist in the current mode.

By default get_interface will return a static mapping when it can't find an active config entry meaning that eventually get_interface will need an update routine to pull out of local running config

The lack of this update routine should be acceptable for now as we have no GUI for users to change the mapping.