Add channels that are within normal 5.8GHz unit range for Rocket M3 usage.
M3 is a 5.8GHz unit with a 2GHz offset transverter.

5MHz width is recomended on these channels.

Send to spectrum analysis for verification that output looks clean.

Stepping stone to working on the wider width of the full 3.4GHz band which has more channels but unlike channels 95-100 (added in this commit) the rest are outside the normal 5.8GHz cal range (but are between two cal points so may be fine)