hardware: Expand board_id routine to allow TPLink devices to be obtained
When a device does not have a subsystem device ID we will fall back to the board_detect routine.

In the case of a TPLink CPE510 the model will be along the lines of "TP-Link CPE510 v1.0"

In the case of TPLink CPE510 this also includes a hardware version so should the hardware be changed in the future we will be able to key off that as well.