bugfix:  Should not keep /etc/config/* on sysupgrade
We don't really need any files in there,  and keeping the folder can interfere with the programs that populate /etc/aredn_include on firstboot from uci-default.

This will cause the node to boot up in default mode (with an AP on its interface on the default channels) during its first boot, and than very quickly reboot and come up under the mesh.