bugfix: Services were not lined up correctly on mesh status page.
Correct services to be correctly lined up.

Note:
It looks like we need to split the two tables eventually on the mesh status page as the services are pushed to the right.

Saving this however to be done in a future release as we may be doing a revamp of this page for sortable fields in the near future and this would fit in with that.