bugfix: UI allows enabling mesh_gw when wan disabled or nat mode
Mesh nodes would need a significant amount of configuration
including some operational checks everytime an interface comes up
to be able to handle this.

The more sane method is to only allow a node to be a mesh_gw when
the WAN port is utilized and the node is in DIRECT mode.

This changes the UI to prevent this setup from being allowed.

fixes AREDN->ticket:184

Change-Id: I1a1289bc914a349149f5842812dbebc0a70442f7