bugfix: Make get_model return model of device instead of board type
Fixes the fact that get_model has done the same as other functions.

This probably always should of been a real model of active device
not just of the hardware type from the begining.

DEV NOTE:
This is in perl, even though we intended to migrate to LUA the
core library that contains this data has yet to be migrated
so we need to add this as a perl program.

Change-Id: Ib8a4ea6dd881c32e6d7660607576aeb5218b1e7a