bugfix: Increase available memory during upgrade/install process
It is possible for the system to run out of memory when dealing
with large file uploads and installs.

As part of the upgrade procedure shutdown services that are not
essential for node operations to allow more memory for install
to take place.

Includes changes to linkled to indicate this new state as it will
be shutdown as part of the cleanup process.

Memory gain (approximate)

dropbear 100kb
linkled  200kb
logd     200kb
odhcp    100kb
snmpd    500kb
xinetd   100kb

Total(approximate): 1200kb (around %4 on 32mb devices)

This is somewhat similar to files/usr/local/bin/upgrade_kill_prep
except that it kills only a select group of services
so that the system can handle the file upload while
upgrade_kill_prep does the final system cleanup to run the full
upgrade.

ref AREDN->ticket:204

Change-Id: Ic6d3aa028725064a97c4723f6d9b36e1e51d87a7