bugfix: DHCP Reservation allows duplicate hostname usage.
The DHCP Name reservation on the ports page would allow
a user to use the name of the mesh node as a hostname
for a reserved host which would mess up the remote display
of hosts on the mesh status screen.

In addition while testing the verification code it was found
that it was case sensitive while dns is case insensitive.

fixes AREDN->ticket:198

Change-Id: Ie51a594f733eca38e8547faeab0260d7b23b8818