aredn: enable nano-m-xw switch to be a dumb switch
Config the NanoStation M2/M5 XW devices to be fully
functional with a standard aredn port definition.
The vlans are not configured in the switch, rather
acting like a dumb swtich, enabling tagged and untagged
packets to exist on the same switch port at the same time.
This commit enables the ports on the switch to directly
send all traffic to all other ports when not configured
directly in vlan mode.