Fix IBSS problem on 2.4GHz (#608)
There's a bug where 2.4GHz channels will associate incorrectly with other 2.4 channels
even though they are on different frequencies. The problem appears to be triggered during
scans when scan probes are sent and received on many different frequencies, but are then
reported as as they arrive as if they come from the selected frequency of the device, causing
incorrect stations to be created. This can only occur for IBSS (adhoc) mode, when the
IBSS frequency is fixed, and when the SSID is also fixed. We do all this for AREDN.
Fortunately these probe contains (but *only* on 2.4 GHz) the actual frequency of the sending
device, so we can avoid incorrectly creating stations which are not really a match.
This situation does not seem to happen on 5GHz channels.
e Please enter the commit message for your changes. Lines starting