Update SUPPORTED_DEVICES.md
Started testing of Mikrotik RBDiscG-5acD. Documentation available via the FCCID suggests this is the exact same board as the RBLHGG-5acD, with a 21db antenna instead of a 24.5db dish. https://fccid.io/TV7LHG5ACD. I have this firmware loaded on 2 nodes of this model, no unsupported banner is showing, and I have them meshed on 131 at 20MHz, testing across the room, with TX power set to 1dbmV on both, reading 100/100/144.4Mbps on the status page.