Periodically sync time if continuous ntpd is unsynchronized
When update-clock is run periodically, if ntpd is running continuously,
check whether it is actually in sync, as measured by a reported stratum
of less than 16.

If it's not in sync, run the full manual clock sync algorithm otherwise
used for our periodic-only synchronization.

This should help with hosts that expect to synchronize from an NTP server
on the internet, for example, when they've temporarily lost the internet
connection. Or alternately, if configured to use a mesh-based NTP server
that has disappeared from the mesh.