Experiemental fix to avoid occasions when the ethernet controller doesnt start up. (#1818)
The old driver used 'at803x-override-sgmii-link-check' to work around a bug where an ethernet
link wouldnt start after it has been auto-negotiated. This flag no longer does anything as this patch
was deprecated as the linux driver has been rewritten.
There is a bug, possibly the same bug as its related to the same startup case, where if the chips goes to sleep
it doesnt not then correct wakeup the RX ethernet MAC. That certainly is the sympton we see where everything
works but no RX traffic is ever received.
Unfortunately there is no way to force the condition to happen, so we dont know if this will fix it until we're sure
we dont see the issue anymore.