Handle missing IPs from tunnels and xlinks. (#2016)
With Babel we treat the endpoint addresses for tunnels and xlinks as
link-local addresses. This means they are no longer distributed as part
of the mesh routing table because they only matter to the nodes at each
end of the link. This causes some trouble for LQM when it comes to monitoring
these links as it no longer has a route to the other end for the endpoint IP.
Fortunately we can use IPv6 to solve this problem, as it has default support
for IPv6 link local addresses, and it's what Babel already uses. We should
utilamately move to entirely IPv6 link local monitoring once OLSR is gone, but
that will take a while and this will, unfortunately, continue to be messy until
then.