Reduce the files we keep in the backups. (#2188)
We added /etc/config file into the backup in an attempt to speed up
the update process by reducing the number of reboots. However, this rarely
works as there's always some change during an update which makes the second
reboot necessary. So, remove the /etc/config files as they serve no purpose.