[BREAKING CHANGE] New format for SSID adhoc halow radios. (#2385)
Regardless of what channel is set in on halow radios in adhoc mode,
they will attempt to join any SSID which matches. This is not what we
want. Include the channel number in the SSID to prevent this. This
is the same thing we do in AP/STA mode for non-halow radios.