Watchdog uses the default timeout rather than setting its own. (#2518)
* Drop the watchdog default time to 300 seconds.
The problem here is that some watchdogs dont honor the time we set but dont error either.
To make it worse when we read the timeout back, they will read back the timeout we set even if they're not
using it. There's not much we can do except use a conservative 'maxWatchdogTimeout' and do the right
thing here in the hope that *some* driver do this correctly.

* As we cannot reliably set the watchdog on all hardware we
we are forced to work with the default value.